Project/video_analzer_use_VLM.py
import cv2
import torch
from transformers import AutoModelWithLMHead, AutoTokenizer

class VideoContentSummarizer:
    def __init__(self, model_name='Sci-fi-vy/Qwen2.5-Omni-7B-GGUF'):
        # Load the model and tokenizer; change this based on your specific model type
        self.tokenizer = AutoTokenizer.from_pretrained(model_name, use_fast=True)
        self.model = AutoModelWithLMHead.from_pretrained(model_name)
        self.device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
        self.model.to(self.device)
        
        # Load additional models for object and action detection here if needed
        # e.g., a pre-trained object detection model and an action recognition model

    def process_video(self, video_path):
        # Open the video file
        cap = cv2.VideoCapture(video_path)
        
        # Process video frames
        frame_count = int(cap.get(cv2.CAP_PROP_FRAME_COUNT))
        frame_rate = int(cap.get(cv2.CAP_PROP_FPS))
        
        frames = []
        for _ in range(min(frame_count, frame_rate * 120)):  # Max 2 min video processing
            ret, frame = cap.read()
            if not ret:
                break
            frames.append(frame)
        
        cap.release()

        # Detect objects and actions in frames
        objects, actions = self.detect_objects_and_actions(frames)

        # Generate a summary
        summary = self.generate_summary(objects, actions)

        return summary

    def detect_objects_and_actions(self, frames):
        objects, actions = [], []
        for frame in frames:
            # Example pseudo function calls; replace with actual logic using a trained model
            detected_objects = self.detect_objects(frame)
            detected_actions = self.detect_actions(frame)
            
            objects.append(detected_objects)
            actions.append(detected_actions)

        return objects, actions

    def detect_objects(self, frame):
        # Placeholder function for object detection
        return ["object1", "object2"]

    def detect_actions(self, frame):
        # Placeholder function for action recognition
        return ["action1", "action2"]

    def generate_summary(self, objects, actions):
        # Create a textual summary from objects and actions using language model
        input_text = f"Objects: {', '.join(set(sum(objects, [])))}. Actions: {', '.join(set(sum(actions, [])))}."
        inputs = self.tokenizer.encode(input_text, return_tensors='pt').to(self.device)
        
        summary_ids = self.model.generate(inputs, max_length=50, num_beams=5, early_stopping=True)
        summary = self.tokenizer.decode(summary_ids[0], skip_special_tokens=True)

        return summary

# Usage
video_summarizer = VideoContentSummarizer()
summary = video_summarizer.process_video("input_video.mp4")
print(summary)