requirements.txt
moviepy==2.2.1
pydantic==2.8.2
numpy>=1.25.0
imageio>=2.5
decorator>=4.0.2
proglog<=1.0.0
python-dotenv>=0.10
pillow>=9.2.0

# Audio processing and AI
transformers>=4.35.0
torch>=2.0.0
librosa>=0.10.0
openai>=1.0.0

# Optional for audio processing
soundfile>=0.12.0

# Video analysis and computer vision
opencv-python>=4.8.0

# LangChain and LangGraph for agent framework
langchain>=0.1.0
langchain-openai>=0.1.0
langgraph>=0.1.0
langchain-core>=0.1.0

# Web API framework
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# Additional dependencies for the agent
asyncio-extras>=1.3.0
uuid
opencv-contrib-python>=4.8.0

# Additional AI models (optional)
# For better video analysis, you can install:
# pip install timm  # For advanced vision models
# pip install ultralytics  # For YOLO object detection
